requirements.txt
streamlit>=1.28.0
yt-dlp>=2023.12.30
openai-whisper>=20231117
pathlib
requests 